requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==1.11.1
charset-normalizer==3.2.0
discord==2.3.1
discord.py==2.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
spotify==0.10.2
yarl==1.9.2
yt-dlp==2024.5.19
